Add a binary 'code' field to the Morse struct
The international morse code standards specify that a dot is 1 tick on,
a dash is 3 ticks on, and the space between letter is 1 tick off. If we
treat each tick on as a 1 and each tick off as a 0, then we can
transform morse code representation into binary. That binary value could
then be used conditionally to light up the LED with each bit.

+++ include/bt-receive/morse.h
@@ -14,7 +14,7 @@
 struct morseStruct {
     char character;
     char convStr[MORSE_MAX_LEN + 1];
-    uint16_t code;
+    uint32_t code;
 };
 
 typedef struct morseStruct * Morse;
